Use test images to create comic archives
Able to test archive read contents
