+++ tests/test_darkseid_comicarchive.py
@@ -1,16 +1,18 @@
 """ Tests for comic archive files """
 import os
-import zipfile
 from pathlib import Path
-from typing import List
 
-import py7zr
 import pytest
 
 from darkseid.comicarchive import ComicArchive
 from darkseid.genericmetadata import GenericMetadata
+from tests.conftest import IMG_DIR
 
 CONTENT = "content"
+
+
+PAGE_TMPL = str(IMG_DIR / "CaptainScience#1_{page_num}.jpg")
+PAGE_THIRTY_SIX = PAGE_TMPL.format(page_num="36")
 
 
 @pytest.fixture()
@@ -22,51 +24,6 @@ def test_metadata() -> GenericMetadata:
     meta_data.notes = "Test comment"
     meta_data.volume = "1"
     return meta_data
-
-
-@pytest.fixture()
-def fake_cbz(tmp_path: Path) -> ComicArchive:
-    image_list: List[Path] = []
-    fake_images: List[str] = [
-        "MarvelTeam-Up058-00fc.jpg",
-        "MarvelTeam-Up058-01.jpg",
-        "MarvelTeam-Up058-02.jpg",
-        "MarvelTeam-Up058-LP.jpg",
-    ]
-    for fake in fake_images:
-        img: Path = tmp_path / fake
-        img.write_text(CONTENT)
-        image_list.append(img)
-
-    z_file: Path = tmp_path / "Aquaman v1 #001 (of 08) (1994).cbz"
-    with zipfile.ZipFile(z_file, "w") as zf:
-        for i in image_list:
-            zf.write(i)
-
-    return ComicArchive(z_file)
-
-
-@pytest.fixture()
-def fake_cb7(tmp_path: Path) -> ComicArchive:
-    image_list: List[Path] = []
-    fake_images: List[str] = [
-        "MarvelTeam-Up058-00fc.jpg",
-        "MarvelTeam-Up058-01.jpg",
-        "MarvelTeam-Up058-02.jpg",
-        "MarvelTeam-Up058-03.jpg",
-        "MarvelTeam-Up058-LP.jpg",
-    ]
-    for fake in fake_images:
-        img: Path = tmp_path / fake
-        img.write_text(CONTENT)
-        image_list.append(img)
-
-    z_file: Path = tmp_path / "Foo Bar v1 #001 (2000).cb7"
-    with py7zr.SevenZipFile(z_file, "w") as archive:
-        for i in image_list:
-            archive.write(i)
-
-    return ComicArchive(z_file)
 
 
 def test_cb7_file_exists(fake_cb7: ComicArchive) -> None:
@@ -126,16 +83,18 @@ def test_removing_metadata_on_cb7_wo_metadata(fake_cb7: ComicArchive) -> None:
     assert fake_cb7.remove_metadata() is True
 
 
-# def test_cb7_get_page(fake_cb7: ComicArchive) -> None:
-#     """Test to set if a page from a comic archive can be retrieved"""
-#     # Get page 2
-#     assert fake_cb7.get_page(0) is not None
+def test_cb7_get_random_page(fake_cb7: ComicArchive) -> None:
+    """Test to set if a page from a comic archive can be retrieved"""
+    page = fake_cb7.get_page(4)
+    with open(PAGE_THIRTY_SIX, "rb") as cif:
+        image = cif.read()
+    assert image == page
 
 
 def test_cb7_metadata_from_filename(fake_cb7: ComicArchive) -> None:
     """Test to get metadata from comic archives filename"""
     test_md = fake_cb7.metadata_from_filename()
-    assert test_md.series == "Foo Bar"
+    assert test_md.series == "Captain Science"
     assert test_md.issue == "1"
     assert test_md.volume == "1"
     assert test_md.year == "2000"
@@ -176,7 +135,7 @@ def test_whether_text_file_is_comic_archive(tmp_path: Path) -> None:
 
 def test_archive_number_of_pages(fake_cbz: ComicArchive) -> None:
     """Test to determine number of pages in a comic archive"""
-    assert fake_cbz.get_number_of_pages() == 4
+    assert fake_cbz.get_number_of_pages() == 5
 
 
 def test_archive_is_writable(fake_cbz: ComicArchive) -> None:
@@ -221,10 +180,12 @@ def test_removing_metadata_on_comic_wo_metadata(fake_cbz: ComicArchive) -> None:
     assert fake_cbz.remove_metadata() is True
 
 
-def test_archive_get_page(fake_cbz: ComicArchive) -> None:
+def test_cbz_get_random_page(fake_cbz: ComicArchive) -> None:
     """Test to set if a page from a comic archive can be retrieved"""
-    # Get page 2
-    assert fake_cbz.get_page(1) is not None
+    page = fake_cbz.get_page(4)
+    with open(PAGE_THIRTY_SIX, "rb") as cif:
+        image = cif.read()
+    assert image == page
 
 
 def test_archive_metadata_from_filename(fake_cbz: ComicArchive) -> None:
@@ -241,7 +202,7 @@ def test_archive_apply_file_info_to_metadata(fake_cbz: ComicArchive) -> None:
     test_md = GenericMetadata()
     fake_cbz.apply_archive_info_to_metadata(test_md)
     # TODO: Need to test calculate page sizes
-    assert test_md.page_count == "4"
+    assert test_md.page_count == "5"
 
 
 def test_archive_export_to_cb7(tmp_path, fake_cbz: ComicArchive) -> None:
@@ -249,4 +210,4 @@ def test_archive_export_to_cb7(tmp_path, fake_cbz: ComicArchive) -> None:
     assert fake_cbz.export_as_cb7(fn) is True
     ca = ComicArchive(fn)
     assert ca.is_sevenzip() is True
-    assert ca.get_number_of_pages() == 4
+    assert ca.get_number_of_pages() == 5
